% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/semipartail.R
\name{hill}
\alias{hill}
\title{Hill estimate of the tail index gamma as a function of threshold t}
\usage{
hill(x, t)
}
\arguments{
\item{x}{the raw data}

\item{t}{the threshold above which to compute the estimator}
}
\value{
a tail index estimate
}
\description{
Hill estimate of the tail index gamma as a function of threshold t
}

