% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/semipartail.R
\name{pareto.neg.loglik}
\alias{pareto.neg.loglik}
\title{Evaluate GPD negative log-likelihood}
\usage{
pareto.neg.loglik(log.sigma, tail.x, gamma)
}
\arguments{
\item{log.sigma}{log scale parameter}

\item{tail.x}{the tail data to be fit}

\item{gamma}{tail index of the GPD}
}
\value{
The negative log-likelihood.
}
\description{
Evaluate GPD negative log-likelihood
}

