% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/semiparmean.R
\name{semipar.mean}
\alias{semipar.mean}
\title{Semiparametric estimate of the expectation of the main population}
\usage{
semipar.mean(main.sample, background.sample, threshold = NULL)
}
\arguments{
\item{main.sample}{the sample of interest}

\item{background.sample}{a (usually large) background sample
used to stabilize tail inference}

\item{threshold}{threshold defining the beginning of the tail
       (selected automatically if not specified)}
}
\value{
Numeric estimate of the expectation of the main sample
        of interest.
}
\description{
Semiparametric estimate of the expectation of the main population
}

