% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/semipartail.R
\name{semipar.tail}
\alias{semipar.tail}
\title{Semiparametric estimate of the main.sample}
\usage{
semipar.tail(main.sample, background.sample, threshold = NULL)
}
\arguments{
\item{main.sample}{the sample of interest}

\item{background.sample}{a (usually large) background sample
used to stabilize tail inference}

\item{threshold}{threshold defining the beginning of
       the tail (selected automatically if not specified)}
}
\value{
An estimated distribution, formatted as pairs
        (X = sample location, weights = amount of probability mass at X),
        where the weights sum to 1.
}
\description{
Semiparametric estimate of the main.sample
}

