% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/semiparquantile.R
\name{semipar.quantile}
\alias{semipar.quantile}
\title{Semiparametric estimate of quantiles of the main population}
\usage{
semipar.quantile(main.sample, background.sample, probs, threshold = NULL)
}
\arguments{
\item{main.sample}{the sample of interest}

\item{background.sample}{a (usually large) background sample used to stabilize tail inference}

\item{probs}{numeric vector of probabilities with values in [0,1].}

\item{threshold}{threshold defining the beginning of the tail (selected automatically if not specified)}
}
\value{
plug-in estimates of the quantiles
}
\description{
Semiparametric estimate of quantiles of the main population
}

